todo

+++ src/main/java/com/sublimee/boot/locks/repository/card/CardRepository.java
@@ -3,6 +3,7 @@ package com.sublimee.boot.locks.repository.card;
 import com.sublimee.boot.locks.model.card.Card;
 
 import javax.persistence.EntityManager;
+import java.math.BigInteger;
 import java.util.UUID;
 import java.util.function.Consumer;
 
@@ -11,7 +12,7 @@ public interface CardRepository {
 
     void persist(Card item);
 
-    Integer count();
+    BigInteger count();
 
     Card find(UUID id);
